{{cookiecutter.project_name}}/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "{{ cookiecutter.project_name }}"
description = "{{ cookiecutter.project_description }}"
version = "0.1.0"
license = "{{ cookiecutter.license|upper }}"

authors = ["{{ cookiecutter.organization }}"]

readme = "README.md"

repository = "https://github.com/{{ cookiecutter.organization }}/{{ cookiecutter.project_name }}"

keywords = []

[tool.nitpick]
style = "https://raw.githubusercontent.com/PyNodeLine/nodeline-python-package/main/styles/nitpick-style-nodeline.toml"

[tool.poetry.dependencies]
python = ">=2.7,>=3.8,<3.10"

[tool.poetry.dev-dependencies]
toml = "^0.10.2"
doit = "^0.36.0"
nox = "^2022.1.7"
pyperclip = "^1.8.2"
python-dotenv = "^0.20.0"
tomlkit = "^0.10.2"
